pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["Abhi16012004 <bigleu88@gmail.com>"]
readme = "README.md"
packages = [
    { include = "airflow" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
numpy = ">=2.3.0,<3.0.0"
scikit-learn = ">=1.7.0,<2.0.0"
fuzzywuzzy = ">=0.18.0,<1.0.0"
langchain = ">=0.3.25,<1.0.0"
langchain-community = ">=0.3.24,<1.0.0"
langchain-huggingface = ">=0.2.0,<1.0.0"
apache-airflow = "^3.0.2"
airflow-python-sdk = "^1.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"